requirements.txt
Django==4.2.7
# mysqlclient==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
